backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
firebase-admin==6.5.0
python-multipart==0.0.9
